harmonize float and ints to their 64bits form

+++ services/backtests/pkg/models/account/account.go
@@ -30,9 +30,9 @@ func (a Account) Validate() error {
 }
 
 func (a Account) ToProtoBuf() *proto.Account {
-	assets := make(map[string]float32)
+	assets := make(map[string]float64)
 	for n, a := range a.Balances {
-		assets[n] = float32(a)
+		assets[n] = float64(a)
 	}
 
 	return &proto.Account{
