+++ services/backtests/pkg/models/account/account_test.go
@@ -50,7 +50,7 @@ func (suite *AccountSuite) TestValidateWithInvalidValue() {
 
 func (suite *AccountSuite) TestFromProtoBuf() {
 	a := FromProtoBuf(&proto.Account{
-		Assets: map[string]float32{
+		Assets: map[string]float64{
 			"asset": 32,
 		},
 	})
@@ -68,5 +68,5 @@ func (suite *AccountSuite) TestToProtoBuf() {
 
 	pb := a.ToProtoBuf()
 	suite.Require().Len(pb.Assets, 1)
-	suite.Require().Equal(float32(32), pb.Assets["asset"])
+	suite.Require().Equal(float64(32), pb.Assets["asset"])
 }
